// web/lib/api/generated/model/fulfilmentsListStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */

export type FulfilmentsListStatus = typeof FulfilmentsListStatus[keyof typeof FulfilmentsListStatus];


export const FulfilmentsListStatus = {
  ALLOCATED: 'ALLOCATED',
  CANCELLED: 'CANCELLED',
  DELIVERED: 'DELIVERED',
  DISPATCHED: 'DISPATCHED',
  PACKED: 'PACKED',
} as const;